mq: patch naming shortcuts
This adds some more options to the mq lookup routine.  It allows
you to use shortcuts in naming patches to push or pop.  You
can now use:

1) a full patch name
2) a number to indicate an offset in the series file
3) a unique substring of the patch name
4) patchname[-+]num to indicate an offset from a given patch.

For case #3 substrings are allowed.  qtip and qbase are aliases for
the top and bottom of the applied patch queue, unless they are already
in the series file as patch names.