Only set mode of new patch if the target file was removed before.
If the file is writable by the user, but owned by a different user, the
chmod will otherwise fail with "Operation not permitted".

Additionally make very sure that the file is only written if either the number
of links is <= 1 or the file was successfully removed.

Maybe this minimal COW code should be replaced by something from util.