doc: use our own rst2man.py script (issue1746)
The rst2man tool has not yet been part of an official Docutils
release, and it is not present in most distributions. This poses a
problem for people who want to install Mercurial from source, or who
want to create a Mercurial package for such a distribution -- how to
specify the build-dependencies?

By including the rst2man.py script with Mercurial people only need a
normal Docutils installation in order to install Mercurial.