Fix --blacklist when --jobs > 1 in run_tests.py.
The options parsing logic replaces the options.blacklist initial value,
a list of filenames, with a dict of test names to filenames.  When
runchildren rebuilds the command line to launch child processes the dict
is flattened and provided as a malformed argument when a filename is
expected.

Fix is to remove the blacklist option from child invocations since the
filtering is handled in the parent anyway.