test-largefiles: partially adapt for Windows
The adaption is partial, because "serve" is not yet officially supported in
tests on Windows.

The test passes on Windows with an experimental testbed that supports
hg serve in tests on Windows - except for some USERCACHE issues.

The added (glob)'s are needed because of backslash <-> shlash issues in paths.